html/rendering/replaced-elements/attributes-for-embedded-content-and-images/picture-aspect-ratio.html

<!doctype html>
<title>Image width and height attributes are used to infer aspect-ratio</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  img {
    width: 100%;
    max-width: 100px;
    height: auto;
  }
</style>
<picture>
  <source srcset="/images/green-100x50.png"></source>
  <img>
</picture>

<picture>
  <source srcset="/images/green-100x50.png" width="100" height="100"></source>
  <img>
</picture>

<picture>
  <source srcset="/images/green-100x50.png" width="100" height="100"></source>
  <img width="50" height="100">
</picture>

<script>
let t = async_test("source width and height attributes are used to infer aspect-ratio in <picture>");
function assert_ratio(img, expected, description) {
  let epsilon = 0.001;
  assert_approx_equals(parseFloat(getComputedStyle(img).width, 10) / parseFloat(getComputedStyle(img).height, 10),
                       expected, epsilon, description);
}

function createPicture(width, height) {
  var picture = document.createElement("picture");
  var source = document.createElement("source");
  if (width !== undefined)
    source.setAttribute("width", width);
  if (height !== undefined)
    source.setAttribute("height", height);
  source.setAttribute("srcset", "/images/green.png");
  picture.appendChild(source);
  var img = document.createElement("img");
  picture.appendChild(img);
  document.body.appendChild(picture);
  return img;
}

function assert_cs(img, val) {
  assert_equals(getComputedStyle(img).aspectRatio, val);
}

// Create and append a new image and immediately check the ratio.
// This is not racy because the spec requires the user agent to queue a task:
// https://html.spec.whatwg.org/multipage/images.html#updating-the-image-data
t.step(function() {
  var img = createPicture(100, 100);
  assert_ratio(img, 1.0);
  assert_cs(img, "auto 100 / 100");

  // Source width/height should take precedence over img attributes.
  img = createPicture(200, 100);
  img.setAttribute("width", 250);
  img.setAttribute("height", 50);
  assert_ratio(img, 2.0);
  assert_cs(img, "auto 200 / 100");

  // Make sure style gets invalidated correctly when the source gets removed.
  img.parentNode.removeChild(img.previousSibling);
  assert_cs(img, "auto 250 / 50");
  img.src = "/images/green.png";
  assert_ratio(img, 5.0);

  img = createPicture(100, undefined);
  img.setAttribute("width", 200);
  img.setAttribute("height", 100);
  assert_ratio(img, 2.0);
  assert_cs(img, "auto 200 / 100");

  // Testing dynamic changes
  img = createPicture(100, 100);
  assert_cs(img, "auto 100 / 100");
  img.previousSibling.setAttribute("height", "300");
  assert_cs(img, "auto 100 / 300");
  img.previousSibling.setAttribute("width", "10");
  assert_cs(img, "auto 10 / 300");
});

onload = t.step_func_done(function() {
  let images = document.querySelectorAll("img");
  assert_ratio(images[0], 2.0, "2.0 is the original aspect ratio of green-100x50.png");
  assert_ratio(images[1], 2.0, "Loaded image's aspect ratio, at least by default, overrides width / height ratio.");
  assert_ratio(images[2], 2.0, "Loaded image's aspect ratio, at least by default, overrides width / height ratio (2).");
});
</script>
